Detector/requirements.txt
fastapi==0.95.1
httpx==0.24.0
jose==1.0.0
numpy==1.24.3
omegaconf==2.3.0
opencv_python==4.7.0.72
Pillow==9.5.0
psutil==5.9.5
pydantic==1.10.7
starlette==0.26.1
torch==1.13.0
torchvision==0.14.0
ultralytics==8.0.88
uvicorn==0.21.1
hydra-core==1.3.2
python-multipart==0.0.6
